requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.1
streamlit>=1.32.0
